# dushengcdn_server/docs/swagger.yaml
basePath: /
definitions:
...
  service.NodeInput:
    properties:
      auto_update_enabled:
        type: boolean
      drain_mode:
        type: boolean
      geo_latitude:
        type: number
      geo_longitude:
        type: number
      geo_manual_override:
        type: boolean
      geo_name:
        type: string
      ip:
        type: string
      name:
        type: string
      pool_name:
        type: string
      public_ips:
        items:
          type: string
        type: array
      scheduling_enabled:
        type: boolean
      tags:
        items:
          type: string
        type: array
      weight:
        type: integer
    type: object
...
info:
  contact: {}
  description: DuShengCDN Server 管理端与 Agent API 文档。
  title: DuShengCDN Server API
  version: "3.0"
paths:
...
  /api/nodes/:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List nodes
      tags:
      - Nodes
    post:
      consumes:
      - application/json
      parameters:
      - description: Node payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/service.NodeInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create node
      tags:
      - Nodes
  /api/nodes/{id}/agent-release:
    get:
      parameters:
      - description: Node ID
        in: path
        name: id
        required: true
        type: integer
      - description: stable or preview
        in: query
        name: channel
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Check latest agent release for node
      tags:
      - Nodes
  /api/nodes/{id}/agent-update:
    post:
      parameters:
      - description: Node ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Request agent self-update on node
      tags:
      - Nodes
  /api/nodes/{id}/delete:
    post:
      parameters:
      - description: Node ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete node
      tags:
      - Nodes
  /api/nodes/{id}/force-sync:
    post:
      parameters:
      - description: Node ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Request force sync config on node
      tags:
      - Nodes
  /api/nodes/{id}/observability:
    get:
      parameters:
      - description: Node ID
        in: path
        name: id
        required: true
        type: integer
      - description: Lookback window in hours
        in: query
        name: hours
        type: integer
      - description: Max records per section
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get node observability details
      tags:
      - Nodes
  /api/nodes/{id}/observability/cleanup:
    post:
      parameters:
      - description: Node ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Cleanup node health events
      tags:
      - Nodes
  /api/nodes/{id}/openresty-restart:
    post:
      parameters:
      - description: Node ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Request openresty restart on node
      tags:
      - Nodes
  /api/nodes/{id}/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: Node ID
        in: path
        name: id
        required: true
        type: integer
      - description: Node payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/service.NodeInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update node
      tags:
      - Nodes
  /api/nodes/bootstrap-token:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get global discovery token
      tags:
      - Nodes
  /api/nodes/bootstrap-token/rotate:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Rotate global discovery token
      tags:
      - Nodes
...
schemes:
- http
- https
securityDefinitions:
  AgentTokenAuth:
    description: Agent API 使用节点专属 Agent Token 或全局 Discovery Token
    in: header
    name: X-Agent-Token
    type: apiKey
  BearerAuth:
    description: 仅部分兼容接口支持长期 Bearer 用户 Token；高危管理接口要求浏览器 session cookie 和 CSRF Token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"